Fix TestInstance_Lifecycle_PER_METHOD_REPETITION, which suffered a regression
after some general improvements on prepare/cleanUp lifecycle.
The issue was how this extension created a new test-instance fresh when
requested by Jupiter pre-destroy callbacks during cleanUp that was
initially suspended and then executes later, after all repetitions are
completed.
The fix here is to leave a mark on repetition parent scope, so that it
will be detected when execution falls back to this parent scope - and
therewith out-of-scope creation of test-instances can be avoided.