Have test use inline-mocking for monitoring on a final class.
A final class of concern here is "UniqueId", which will be final on
JUnit Platform version 6.x. A fairly new Mockito version is used when
testing JUnit-6 compatibility but there are regression tests suites on
JUnit-5 using older Mockito releases that don't support inline-mocking.
Therefore default behavior is to use default mock-factory and only
apply inlining when necessary on JUnit-6.
(To always use inlining mocking for all classes on JUnit-6 would not
be very good, however, because it would also cause mocking of some
static methods that would alter the test scenario if mocked.)