Shape up pom by enforcing separate JUnit-dependency versions between scopes compile and test.
The new shape makes no technical difference to projects that consume
the artifact, because the effective JUnit compile-dependencies of
concern are still "optional", so a consumer project is still obliged
to declare its own JUnit dependencies.

The separation of JUnit versions hardwires the dependencies during
compilation of the main sources to just the very few specific 5.x
JUnit-release artifacts that are needed to compile the main sources,
whereas the tests are compiled and executed with JUnit dependency
versions that unfold from the "version.junit5or6" and "version.platform"
properties, for which specific versions are specified by the
"junit_<version>" profiles. But their non-profile default values are
version-ranges that unfold as test dependencies on JUnit-6' most
current (possibly milestone or RC) release.

This means that tests are expected to be compiled with a classpath
that includes JUnit-6, which requires JDK-17 or later. Therefore
maven-compiler-plugin is reconfigured to use JDK-17 for compiling
tests that run with JUnit release 5.x or 6.x - and that covers most
of the tests! The JDK-17 compilation will still produce java-8
classes, as before, so the tests can also run with JUnit-5.x on
JDK-8 or later. However, the compiler goal "compile-jdk6-tests"
still uses JDK-8 to compile a few JUnit-4 tests that are expected
to pass on all JDKs starting from relese 6 and later.

For a LazyParams maintainer, this means both JDK-8 and JDK-17 must be
configured by Maven toolchain to build this project, where JDK-8 is
used to compile the main sources and a few JUnit-4 tests - and JDK-17
is used to compile most of the tests. Default for running the tests is
JDK-17, which supports all JUnit versions - but the maintainer still
has test profiles for verifying test-execution on other JDKs with
specific versions of JUnit and ByteBuddy.

The need for these two JDKs on toolchain might appear as a complicated
dilemma for the next LazyParams maintainer but it is most likely a good
move that will make development much easier in the long run. A fact
is that regression tests on different JDKs have always been part of
development for this project but the previous pom structure did not
provide very good guidance on how to go about - e.g. the JUnit versions
for compiling the main sources were provided by a goofy kind-of-default
activation of special profile "default build" and some tricky
bootstrapping would be necessary to compile any test on more recent
JUnit releases, while preferably making sure no dependency on this
newer JUnit release is leaked to the main compilation. But now the
main compilation has its own dependencies and it will take place on
JDK-8 - independently of whichever JUnit release has been staged for
the test compilation, which will happen on JDK-17 that is compatible
with all JUnit releases and will also be default for executing the tests.