Improve security a bit by using pgpverify-maven-plugin
to check signatures on plugins and dependencies.

Of special concern are the JUnit dependencies version ranges, which help
trigger download and test with latest JUnit release, as soon as a new
release (possibly milestone or release-candidate) is available. This can
hopefully help the next developer feel confident that newly downloaded
JUnit artifacts are indeed valid, by automatically checking that they
have valid signatures and that these signatures are made with keys that
are controlled by recognized JUnit maintainers.

On top of just checking the dependencies we also try to make the most
out of this by also checking plugins and their dependencies.