Stall initialization of field AdviceFor#methodInterceptor until actually being used
by an implementation that relies on method #on() for building #methods2advice.
This can improve performance a little because #methodInterceptor will only be
created on implementations that actually use it.
Also make some refactoring that seems appropriate with this modification,
such as replacing variable "tClass" with field "templateClass", which is set as null if generic template-type is Object.