Have executions of maven-toolchains-plugin and jacoco-maven-plugin postboned,
because their executions are of no concern until the Failsafe and
Surefire test executions. Therefore the executions of concerned are
moved to phase "process-test-classes" that takes place right before
phase "test".
Also, to better reflect their role in the overall build lifecycle, the
plugin XML configuration are moved to right before configuration of the
test-execution plugins Failsafe and Surefire.